+++ python/dalex/dalex/dataset_level/__init__.py
@@ -1,9 +1,11 @@
 from ._aggregated_profiles.object import AggregatedProfiles
 from ._model_performance.object import ModelPerformance
 from ._variable_importance.object import VariableImportance
+from ._residual_diagnostics import ResidualDiagnostics
 
 __all__ = [
     "ModelPerformance",
     "VariableImportance",
-    "AggregatedProfiles"
+    "AggregatedProfiles",
+    "ResidualDiagnostics"
 ]