+++ python/dalex/dalex/dataset_level/_residual_diagnostics/__init__.py
@@ -1,0 +1,5 @@
+from .object import ResidualDiagnostics
+
+__all__ = [
+    "ResidualDiagnostics"
+]